+++ test/java/src/org/qore/jni/test/ClosureTest1.java
@@ -3,7 +3,7 @@ package org.qore.jni.test;
 import org.qore.jni.*;
 
 public class ClosureTest1 implements QoreClosureMarker {
-    public String call(int i) {
+    public String call(long i) {
         return "closure-" + String.valueOf(i);
     }
 }
