+++ test/java/src/org/qore/jni/test/ClosureTest2.java
@@ -3,6 +3,6 @@ package org.qore.jni.test;
 import org.qore.jni.*;
 
 public class ClosureTest2 implements QoreClosureMarker {
-    public void call(int i) {
+    public void call(long i) {
     }
 }
